\name{getCandidateInteractions}
\alias{getCandidateInteractions}
\title{
Get Candidate Interactions
}
\description{
This function will indicate interactions that are most probably causing the difference between samples.
}
\usage{
getCandidateInteractions(output, peakFiles, defchic.settings, pcol = "weighted_padj", method = "min", deltaAsinhScore=1, pvcut = 0.05)
}
\arguments{
  \item{output}{final output of \code{\link{chicdiffPipeline}} (or, equivalently, of \code{\link{IHWcorrection}}, containing information on bait-region interactions.
}
  \item{peakFiles}{
The location of the peak file (or several comma-separated peak files).
}
  \item{defchic.settings}{
The parameter list generated by \code{\link{setChicExperiment}}
}
  \item{pcol}{The name of the column in \code{output} containing the p-values of interest (default: weighted_padj).}
  \item{method}{The method to combine region-level p-values for interactions mapping to multiple overlapping regions (default: min [take the minimum value], alternative method: hmp [use \code{p.hmp} from package \code{harmonicmeanp}).}
  \item{pvcut}{User-defined max p-value for output filtering (default: 0.05)}
  \item{deltaAsinhScore}{User-defined minimal absolute difference between the asinh of Chicago scores of studied samples for output filtering (default: 1).}}

\value{
A data.table with full information about candidate interactions, including chromosome, coordinates, genes, Chicago scores and weighted pvalues.
}
\author{
Valeriya Malysheva and Mikhail Spivakov
}

\seealso{
  \code{\link{setChicExperiment}}, \code{\link{chicdiffPipeline}}
}
\examples{
## path to recover test data
dataPath <- system.file("extdata", package="ChicdiffData")

## read in some raw data and settings
testPeakFiles <- file.path(dataPath,"peakMatrix_cd4_mono_unmerged.txt")
testDefChicSettings <- readRDS(file.path(dataPath, "defChicSettings.Rds"))

#load pre-processed data
resultsPath <- file.path(dataPath, "CD4_Mono_results")
testOutput <- file.path(resultsPath, "Weighted.Rds")

#get candidate interactions
out <- getCandidateInteractions(defchic.settings=testDefChicSettings, output = testOutput, 
                                peakFiles = testPeakFiles, pvcut = 0.05, deltaAsinhScore = 1)
}

